+++ packages/services/package.json
@@ -24,15 +24,13 @@
     "./utils": {
       "import": "./src/utils/index.ts",
       "require": "./src/utils/index.ts"
-    }
-  },
-  "type": "module",
-  "exports": {
+    },
     "./estimate": {
       "import": "./src/estimate/index.ts",
       "require": "./src/estimate/index.ts"
     }
   },
+  "type": "module",
   "scripts": {
     "lint": "eslint . --max-warnings 0",
     "check-types": "tsc --noEmit"
