+++ packages/services/src/estimate/components/estimate-list/empty-state/index.styles.ts
@@ -1,13 +1,14 @@
 import { css } from '@emotion/react';
 import { theme } from '@daengle/design-system';
 
-export const wrapper = css`
+export const wrapper = (hasOptions: boolean) => css`
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
   text-align: center;
   gap: 15px;
+  margin-top: ${hasOptions ? '150px' : '100px'};
 `;
 
 export const emptyButton = css`
